Converted whitelistPending to an object with key:value of domain: time started. If somehow the browser gets completely shut down in the middle of doing the history dive, it should restart from scratch when the browser is next opened.
Made new page tracker only record a new event if the new page is different than the old page.
This is mainly to avoid the issue of requests made by scripts that load ads/info/whatever
triggering the background task and dozens of 'hits' to the current page getting recorded.

Fixed a number of bugs and removed extraneous leftover code.