Popup and background now message each other about various events:
Popup triggers the history crawl when a domain or URL is added via button.
Background tells the popup when any of the store whitelists are updated so that the popup can sync that info.
It also indicates when a history crawl is done so that the popup can re-check for page Content.

The history crawl is now a function outside the webNavigation.onCompleted listener so that the manual trigger from popup can be done.
It's passed the Connection as a parameter since now the Connection isn't known when the function is instantiated.

Popup now has some text indicating when the history crawl for the current page is running.

New Content was missing the provider_id_string.

Fixed a bug with the whitelistPending.
For some reason bogus objects were being written as values instead of dates.
These would reuslt in those domains never being cleared from Pending status.
Now, the expired Pending check also clears entries with non-Date values.